// apps/element-service/src/models/element.model.ts
export type ElementModel = {
  elementId: string;
  projectId: string;
  type: string;
  x: number;
  y: number;
  z: number;
  width: number;
  height: number;
  depth: number;
  color: string;
};

export type ElementInputModel = {
  type: string;
  x: number;
  y: number;
  z: number;
  width: number;
  height: number;
  depth: number;
  color: string;
};